requirements.txt
langchain-community
langchain-openai
openai
faiss-cpu
youtube-transcript-api>=0.6.0
python-dotenv
langchain-core
langchain-text-splitters